Add isolated EasyMagpie serving tests
