+++ .github/workflows/cicd-main-speech.yml
@@ -41,6 +41,58 @@ jobs:
       dockerfile: docker/Dockerfile
       runner: ${{ inputs.runner }}
 
+  easymagpie-vllm-serving-changes:
+    runs-on: ubuntu-latest
+    outputs:
+      run: ${{ steps.decision.outputs.run }}
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v6
+        with:
+          fetch-depth: 0
+      - name: Get PR info
+        id: get-pr-info
+        if: startsWith(github.ref, 'refs/heads/pull-request/')
+        uses: nv-gha-runners/get-pr-info@main
+      - name: Get changed files
+        id: changed-files
+        uses: step-security/changed-files@v45.0.1
+        with:
+          sha: ${{ github.sha }}
+          base_sha: ${{ fromJSON(steps.get-pr-info.outputs.pr-info || '{}').base.sha || github.event.pull_request.base.sha }}
+          files: |
+            .github/workflows/cicd-main-speech.yml
+            examples/tts/easymagpie_vllm_omni/**
+            tests/collections/tts/easymagpie_vllm_omni/serving/**
+            tests/functional_tests/L0_Unit_Tests_GPU_TTS_EasyMagpie_vLLM.sh
+      - name: Decide whether to run serving tests
+        id: decision
+        env:
+          ANY_CHANGED: ${{ steps.changed-files.outputs.any_changed }}
+          EVENT_NAME: ${{ github.event_name }}
+          TESTS_TO_RUN: ${{ inputs.test_to_run }}
+        run: |
+          run="$ANY_CHANGED"
+
+          if [[ "$EVENT_NAME" == "schedule" ]]; then
+            run=false
+          elif [[ "$EVENT_NAME" == "workflow_dispatch" ]] && \
+              jq -e 'index("all") != null or index("L0_Unit_Tests_GPU_TTS_EasyMagpie_vLLM") != null' \
+                <<<"$TESTS_TO_RUN" >/dev/null; then
+            run=true
+          fi
+
+          echo "run=$run" | tee -a "$GITHUB_OUTPUT"
+
+  easymagpie-vllm-serving-build:
+    needs: [easymagpie-vllm-serving-changes]
+    if: needs.easymagpie-vllm-serving-changes.outputs.run == 'true'
+    uses: ./.github/workflows/_build_container.yml
+    with:
+      image-name: nemo_container_easymagpie_vllm
+      dockerfile: examples/tts/easymagpie_vllm_omni/Dockerfile.ci
+      runner: ${{ inputs.runner }}
+
   unit-tests:
     strategy:
       fail-fast: false
@@ -145,6 +197,28 @@ jobs:
           timeout: ${{ matrix.timeout || 10 }}
           cpu-only: ${{ matrix.cpu-only || false }}
           is_optional: ${{ matrix.is-optional || false }}
+
+  easymagpie-vllm-serving-tests:
+    needs: [easymagpie-vllm-serving-build]
+    runs-on: ${{ inputs.runner }}
+    name: L0_Unit_Tests_GPU_TTS_EasyMagpie_vLLM
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v6
+      - name: Checkout NeMo
+        uses: actions/checkout@v6
+        with:
+          path: ${{ github.run_id }}
+      - name: main
+        uses: ./.github/actions/test-template
+        with:
+          runner: ${{ runner.name }}
+          script: L0_Unit_Tests_GPU_TTS_EasyMagpie_vLLM
+          is_unit_test: true
+          tests_to_run: ${{ inputs.test_to_run }}
+          image: ${{ needs.easymagpie-vllm-serving-build.outputs.image }}
+          test-data-path: ${{ vars.DEFAULT_TEST_DATA_PATH || '/mnt/datadrive/TestData' }}
+          timeout: 15
 
   e2e-tests:
     strategy:
