+++ examples/tts/easymagpie_vllm_omni/Dockerfile.ci
@@ -1,4 +1,4 @@
-# Copyright (c) 2020-2025, NVIDIA CORPORATION.
+# Copyright (c) 2026, NVIDIA CORPORATION.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,4 +11,18 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-TORCH_FORCE_NO_WEIGHTS_ONLY_LOAD=1 NEMO_NUMBA_MINVER=0.53 CUDA_VISIBLE_DEVICES=0 coverage run -a --data-file=/workspace/.coverage --source=/workspace/ -m pytest tests/collections/tts -m "not pleasefixme" --ignore=tests/collections/tts/easymagpie_vllm_omni/serving --with_downloads
+
+FROM vllm/vllm-openai:v0.24.0
+
+COPY examples/tts/easymagpie_vllm_omni/requirements.txt /tmp/easymagpie-requirements.txt
+
+RUN python3 -m pip install --no-cache-dir \
+    -r /tmp/easymagpie-requirements.txt \
+    coverage \
+    pytest \
+    pytest-asyncio \
+    "setuptools>=64"
+
+WORKDIR /workspace
+
+ENTRYPOINT []
