+++ tests/collections/tts/easymagpie_vllm_omni/conversion/conftest.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2020-2025, NVIDIA CORPORATION.
+# Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES.  All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,4 +11,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-TORCH_FORCE_NO_WEIGHTS_ONLY_LOAD=1 NEMO_NUMBA_MINVER=0.53 CUDA_VISIBLE_DEVICES=0 coverage run -a --data-file=/workspace/.coverage --source=/workspace/ -m pytest tests/collections/tts -m "not pleasefixme" --ignore=tests/collections/tts/easymagpie_vllm_omni/serving --with_downloads
+"""Path setup for Speech-side EasyMagpie checkpoint conversion tests."""
+from __future__ import annotations
+
+import sys
+from pathlib import Path
+
+EASYMAGPIE_ROOT = Path(__file__).resolve().parents[5] / "examples" / "tts" / "easymagpie_vllm_omni"
+SCRIPTS_DIR = EASYMAGPIE_ROOT / "scripts"
+
+# Conversion uses Speech and pure PyTorch helpers, never the vLLM runtime.
+sys.path.insert(0, str(EASYMAGPIE_ROOT))
+sys.path.insert(0, str(SCRIPTS_DIR))
