+++ tests/collections/tts/easymagpie_vllm_omni/serving/test_audio_output.py
