+++ tests/collections/tts/easymagpie_vllm_omni/serving/test_benchmark_server.py
@@ -13,16 +13,12 @@
 # limitations under the License.
 from __future__ import annotations
 
-import sys
-from pathlib import Path
 
 import pytest
 
 pytest.importorskip("numpy")
 pytest.importorskip("requests")
 
-SCRIPTS_DIR = Path(__file__).parents[1] / "scripts"
-sys.path.insert(0, str(SCRIPTS_DIR))
 import benchmark_server as benchmark  # noqa: E402
