+++ tests/collections/tts/easymagpie_vllm_omni/serving/test_scheduler.py
