+++ tests/functional_tests/L0_Unit_Tests_CPU_TTS.sh
@@ -11,4 +11,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-CUDA_VISIBLE_DEVICES="" NEMO_NUMBA_MINVER=0.53 coverage run -a --data-file=/workspace/.coverage --source=/workspace/ -m pytest tests/collections/tts -m "not pleasefixme" --cpu --with_downloads --relax_numba_compat
+CUDA_VISIBLE_DEVICES="" NEMO_NUMBA_MINVER=0.53 coverage run -a --data-file=/workspace/.coverage --source=/workspace/ -m pytest tests/collections/tts -m "not pleasefixme" --ignore=tests/collections/tts/easymagpie_vllm_omni/serving --cpu --with_downloads --relax_numba_compat
