+++ tests/functional_tests/L0_Unit_Tests_GPU_TTS_EasyMagpie_vLLM.sh
@@ -1,4 +1,6 @@
-# Copyright (c) 2020-2025, NVIDIA CORPORATION.
+#!/usr/bin/env bash
+
+# Copyright (c) 2026, NVIDIA CORPORATION.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,4 +13,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-TORCH_FORCE_NO_WEIGHTS_ONLY_LOAD=1 NEMO_NUMBA_MINVER=0.53 CUDA_VISIBLE_DEVICES=0 coverage run -a --data-file=/workspace/.coverage --source=/workspace/ -m pytest tests/collections/tts -m "not pleasefixme" --ignore=tests/collections/tts/easymagpie_vllm_omni/serving --with_downloads
+
+set -euo pipefail
+
+python3 -m pip install --no-build-isolation --no-deps -e examples/tts/easymagpie_vllm_omni
+
+CUDA_VISIBLE_DEVICES=0 coverage run \
+    -a \
+    --data-file=/workspace/.coverage \
+    --source=/workspace/examples/tts/easymagpie_vllm_omni \
+    -m pytest \
+    --confcutdir=tests/collections/tts/easymagpie_vllm_omni/serving \
+    tests/collections/tts/easymagpie_vllm_omni/serving
